feat(m2.8): extract Discovery domain bounded context with zero functional impact
